\name{iND750}
\docType{data}
\alias{iND750}

\encoding{utf8}

\title{Saccharomyces cerevisiae Metabolic Model iND750}

\description{
  The dataset is a genome scale metabolic network of the \emph{Saccharomyces cerevisiae}. It consists of 1149 internal reactions, 116 exchange
  reactions and a biomass objective function.
}

\usage{data(iND750)}

\format{An object of class \code{modelorg}}

%\source{\url{http://gcrg.ucsd.edu/Downloads/EcoliCore}}

\references{
Duarte NC, Herrgard MJ, Palsson BO: Reconstruction and validation of
Saccharomyces cerevisiae iND750, a fully compartmentalized genomescale
metabolic model. Genome Res 2004, 14(7):1298-1309.
}

\keyword{datasets}
