\name{sybilEFBA-package}
\alias{sybilEFBA-package}
\alias{sybilEFBA}
\docType{package}
\title{
  Use gene expression data to improve Flux Balance Analysis predictions
}
\description{
The package \code{sybilEFBA} implements some ideas to get flux predictions that are better correlated to gene expression data.
}
\details{
\tabular{ll}{
Package: \tab sybilEFBA\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-06-03\cr
License: \tab GPL Version 3\cr
LazyLoad: \tab yes\cr
Depends: \tab \code{\link{sybil}}, methods\cr
}
}
\author{Abdelmoneim Amer Desouki}
\keyword{ package }
\seealso{
	\code{\link{sybil}} 
	\code{\link{eFBA_gene}} 
	\code{\link{eFBA_rxn}} 
	\code{\link{FECorr}} 
}